+++ src/components/MetadataCollections/MetadataCollectionView.test.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import { render, screen } from '@folio/jest-config-stripes/testing-library/react';