+++ src/components/MetadataSources/SourceManagement/Contact/ContactFieldArray.js
@@ -1,5 +1,4 @@
 import _ from 'lodash';
-import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Field } from 'react-final-form';