+++ src/components/MetadataSources/SourceManagement/Contact/DisplayContactLinkContact.js
@@ -1,5 +1,4 @@
 import { isEmpty } from 'lodash';
-import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import { useQuery } from 'react-query';