+++ src/index.test.js
@@ -1,5 +1,4 @@
 import { noop } from 'lodash';
-import React from 'react';
 import { Router } from 'react-router-dom';
 import { createMemoryHistory } from 'history';
 import {