+++ src/routes/CollectionsRoute.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import React, { useRef, useEffect, useState } from 'react';
+import { useRef, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';