+++ test/jest/helpers/withIntlConfiguration.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { IntlProvider } from 'react-intl';
 import stripesComponentsTranslations from '@folio/stripes-components/translations/stripes-components/en';
 import stripesSmartComponentsTranslations from '@folio/stripes-smart-components/translations/stripes-smart-components/en';