+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-finc-config
 
 ## [6.2.0] (IN PROGRESS)
+* React v19: refactor away from default props for functional components([UIFC-355](https://folio-org.atlassian.net/browse/UIFC-355))
 
 ## [6.1.0](https://github.com/folio-org/ui-finc-config/tree/v6.1.0) (2024-03-20)
 * Remove deprecated pane properties ([UIFC-337](https://issues.folio.org/browse/UIFC-337))